This should apparently be an uint

+++ include/stddef.h
@@ -1,5 +1,5 @@
 #pragma once
 
 #define NULL 0x00
-typedef unsigned long size_t;
+typedef unsigned int size_t;
